+++ src/main/java/com/zuniorteam/lotto/LottoGameWebBoard.java
@@ -1,10 +1,5 @@
 package com.zuniorteam.lotto;
 
-import com.zuniorteam.lotto.core.LottoAutoMachine;
-import com.zuniorteam.lotto.core.LottoOffice;
-import com.zuniorteam.lotto.core.LottoSeller;
-import com.zuniorteam.lotto.view.console.InputView;
-import com.zuniorteam.lotto.view.console.OutputView;
 import com.zuniorteam.lotto.web.route.PostBuyLottoRoute;
 import com.zuniorteam.lotto.web.route.PostMatchLottoRoute;
 
@@ -12,15 +7,15 @@ import static spark.Spark.*;
 
 public class LottoGameWebBoard {
     private static final String RESOURCE_DIRECTORY = "/templates";
+    public static final int PORT = 8080;
 
     public void playGame() {
-        port(8080);
+        port(PORT);
 
         staticFiles.location(RESOURCE_DIRECTORY);
 
         post("/buyLotto", new PostBuyLottoRoute());
         post("/matchLotto", new PostMatchLottoRoute());
     }
-
 
 }
