+++ src/main/java/com/zuniorteam/lotto/web/render/CustomTemplateEngine.java
@@ -6,13 +6,13 @@ import spark.template.thymeleaf.ThymeleafTemplateEngine;
 
 import java.util.Map;
 
-public class TemplateEngine {
+public class CustomTemplateEngine {
 
 
-    private TemplateEngine() {
+    private CustomTemplateEngine() {
     }
 
-    private static final spark.TemplateEngine TEMPLATE_ENGINE = new ThymeleafTemplateEngine();
+    private static final TemplateEngine TEMPLATE_ENGINE = new ThymeleafTemplateEngine();
 
     public static String render(Map<String, Object> model, String viewName) {
         return TEMPLATE_ENGINE.render(new ModelAndView(model, viewName));
